requirements.txt
Flask~=3.1.0
torch~=2.4.0
torchvision~=0.19.0
pillow~=10.4.0